reorganize components and change some urls based on back end
